feat: add balance hook and related pages skeletons
