+++ wallet/package.json
@@ -15,6 +15,7 @@
     "@0xsequence/network": "^2.2.4",
     "@0xsequence/waas": "^2.2.4",
     "@react-oauth/google": "^0.12.1",
+    "@tanstack/react-query": "^5.66.0",
     "@walletconnect/sign-client": "^2.18.0",
     "ethers": "6.13.3",
     "framer-motion": "^11.5.2",
@@ -28,6 +29,7 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
+    "@tanstack/eslint-plugin-query": "^5.66.1",
     "@trivago/prettier-plugin-sort-imports": "^4.3.0",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
