+++ wallet/src/context/AuthContext.tsx
@@ -1,6 +1,8 @@
-import React, { createContext, useContext, useEffect, useState } from 'react'
+import { PropsWithChildren, createContext, useContext, useEffect, useState } from 'react'
+import { useNavigate } from 'react-router'
 import { useSnapshot } from 'valtio'
 
+import { ROUTES } from '../routes'
 import { sequenceWaas } from '../waasSetup'
 import { WalletTransport } from '../walletTransport'
 
@@ -17,7 +19,8 @@ interface AuthContextType {
 
 const AuthContext = createContext<AuthContextType | undefined>(undefined)
 
-export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
+export const AuthProvider = ({ children }: PropsWithChildren) => {
+  const navigate = useNavigate()
   const [authState, setAuthState] = useState<AuthState>({ status: 'loading' })
   const walletTransportSnapshot = useSnapshot(walletTransport.state)
 
@@ -40,10 +43,15 @@ export const AuthProvider: React.FC<{ children: React.ReactNode }> = ({ children
   }
 
   const signOut = async () => {
-    await sequenceWaas.dropSession()
-    setAuthState({ status: 'signedOut' })
-    walletTransport.setSignedInState(null)
-    localStorage.clear()
+    try {
+      setAuthState({ status: 'signedOut' })
+      await sequenceWaas.dropSession()
+      walletTransport.setSignedInState(null)
+      localStorage.clear()
+      navigate(ROUTES.AUTH)
+    } catch {
+      setAuthState(authState)
+    }
   }
 
   return (
