+++ wallet/src/main.tsx
@@ -1,9 +1,10 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 
-import { App } from './App.tsx'
+import { App as Router } from './App.tsx'
 import './index.css'
 import { BrowserRouter } from 'react-router'
+import { AppContextProvider } from './context/AppContext.tsx'
 
 const favicon = document.getElementById('favicon')
 const appTitle = document.getElementById('app-title')
@@ -15,11 +16,14 @@ if (favicon instanceof HTMLLinkElement && smallLogo) {
 if (appTitle && projectName) {
   appTitle.textContent = projectName + ' ' + 'Wallet'
 }
-
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <BrowserRouter>
-      <App />
+      <div id="app">
+        <AppContextProvider>
+          <Router />
+        </AppContextProvider>
+      </div>
     </BrowserRouter>
   </StrictMode>
 )
