+++ wallet/src/pages/SendPage.tsx
@@ -1,0 +1,9 @@
+import { Box } from '@0xsequence/design-system'
+
+export const SendPage = () => {
+  return (
+    <Box background="white">
+      <h1>Send Page</h1>
+    </Box>
+  )
+}
