+++ wallet/src/pages/TransactionsPage.tsx
@@ -1,0 +1,12 @@
+
+import { Box } from "@0xsequence/design-system";
+
+export const TransactionsPage: React.FC = () => {
+  return (
+    <Box background="white">
+      <h1>Txns Page</h1>
+      <p>Welcome to the Txns page!</p>
+    </Box>
+  );
+};
+
