+++ wallet/src/routes.ts
@@ -1,16 +1,16 @@
 export const ROUTES = {
-	AUTH: "/auth",
-	HOME: "/",
-	ASSETS: {
-		INDEX: "/assets",
-		COLLECTIBLES: "/assets/collectibles",
-	},
+  AUTH: "auth",
+  HOME: "/",
+  TOKENS: "tokens",
+  COLLECTIBLES: "collectibles",
+  SEND: "send",
+  TRANSACTIONS: "transactions",
 } as const;
 
 type RouteValues<T> = T[keyof T];
 type NestedRouteValues<T> = T extends object
-	? RouteValues<T> | { [K in keyof T]: NestedRouteValues<T[K]> }[keyof T]
-	: never;
+  ? RouteValues<T> | { [K in keyof T]: NestedRouteValues<T[K]> }[keyof T]
+  : never;
 
 export type AppRoute = NestedRouteValues<typeof ROUTES>;
 
@@ -18,5 +18,5 @@ export type PublicRoutes = typeof ROUTES.AUTH;
 export type ProtectedRoutes = Exclude<AppRoute, PublicRoutes>;
 
 export const isPublicRoute = (route: string): route is PublicRoutes => {
-	return route === ROUTES.AUTH;
+  return route === ROUTES.AUTH;
 };
