+++ wallet/src/utils/time.const.ts
@@ -1,0 +1,5 @@
+export const TIME = {
+  SECOND: 1 * 1000,
+  MINUTE: 60 * 1000,
+  HOUR: 60 * 60 * 1000
+}
